# Conflicts:
#	java-vtl-model/src/main/java/no/ssb/vtl/model/VTLObject.java
#	java-vtl-parser/src/main/antlr4/no/ssb/vtl/parser/VTL.g4
#	java-vtl-script/src/main/java/no/ssb/vtl/script/visitors/AssignmentVisitor.java

+++ java-vtl-script/src/main/java/no/ssb/vtl/script/support/VTLPrintStream.java
@@ -67,7 +67,7 @@ public class VTLPrintStream extends PrintStream {
         );
         structure.addRow(ds.values().stream()
                 .map(Component::hashCode)
-                .map(i -> Integer.toString(i))
+                .map(i -> Long.toString(i))
                 .collect(Collectors.toList())
         );
         structure.showTable(this);
